requirements.txt
openai 
python-dotenv 
duckduckgo-search
wikipedia 
langchain-openai
langchain-community 
langchainhub 
langchain-chroma